requirements.txt
annotated-types==0.6.0
anyio==3.7.1
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.104.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
h11==0.14.0
idna==3.4
opentelemetry-api==1.12.0
opentelemetry-exporter-otlp==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-exporter-otlp-proto-http==1.12.0
opentelemetry-proto==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
protobuf==3.20.3
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
email-validator==2.1.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wrapt==1.16.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1
